+++ packages/backend/src/models/SpaceModel.ts
@@ -12,6 +12,7 @@ import {
     ProjectRole,
     Space,
     SpaceDashboard,
+    SpaceMemberRole,
     SpaceQuery,
     SpaceShare,
     SpaceSummary,
@@ -51,8 +52,8 @@ import {
 } from '../database/entities/savedCharts';
 import {
     DbSpace,
-    SpaceShareTableName,
     SpaceTableName,
+    SpaceUserAccessTableName,
 } from '../database/entities/spaces';
 import { UserTableName } from '../database/entities/users';
 import { DbValidationTable } from '../database/entities/validation';
@@ -109,14 +110,14 @@ export class SpaceModel {
                 `${PinnedSpaceTableName}.pinned_list_uuid`,
             )
             .leftJoin(
-                SpaceShareTableName,
-                `${SpaceShareTableName}.space_id`,
-                `${SpaceTableName}.space_id`,
+                SpaceUserAccessTableName,
+                `${SpaceUserAccessTableName}.space_uuid`,
+                `${SpaceTableName}.space_uuid`,
             )
             .leftJoin(
                 'users',
-                `${SpaceShareTableName}.user_id`,
-                `${UserTableName}.user_id`,
+                `${SpaceUserAccessTableName}.user_uuid`,
+                `${UserTableName}.user_uuid`,
             )
             .where((q) => {
                 q.where(`${UserTableName}.user_uuid`, userUuid).orWhere(
@@ -286,7 +287,7 @@ export class SpaceModel {
     async find(filters: {
         projectUuid?: string;
         spaceUuid?: string;
-    }): Promise<SpaceSummary[]> {
+    }): Promise<Omit<SpaceSummary, 'userAccess'>[]> {
         const transaction = Sentry.getCurrentHub()
             ?.getScope()
             ?.getTransaction();
@@ -317,14 +318,14 @@ export class SpaceModel {
                     `${PinnedSpaceTableName}.pinned_list_uuid`,
                 )
                 .leftJoin(
-                    'space_share',
-                    'space_share.space_id',
-                    'spaces.space_id',
+                    `${SpaceUserAccessTableName}`,
+                    `${SpaceUserAccessTableName}.space_uuid`,
+                    'spaces.space_uuid',
                 )
                 .leftJoin(
                     'users as shared_with',
-                    'space_share.user_id',
-                    'shared_with.user_id',
+                    `${SpaceUserAccessTableName}.user_uuid`,
+                    'shared_with.user_uuid',
                 )
                 .groupBy(
                     `${PinnedListTableName}.pinned_list_uuid`,
@@ -611,15 +612,15 @@ export class SpaceModel {
                     );
                 },
             )
-            .leftJoin(SpaceShareTableName, function joinSpaceShareTable() {
+            .leftJoin(SpaceUserAccessTableName, function joinSpaceShareTable() {
                 this.on(
-                    `${UserTableName}.user_id`,
+                    `${UserTableName}.user_uuid`,
                     '=',
-                    `${SpaceShareTableName}.user_id`,
+                    `${SpaceUserAccessTableName}.user_uuid`,
                 ).andOn(
-                    `${SpaceTableName}.space_id`,
+                    `${SpaceTableName}.space_uuid`,
                     '=',
-                    `${SpaceShareTableName}.space_id`,
+                    `${SpaceUserAccessTableName}.space_uuid`,
                 );
             })
             .leftJoin(
@@ -662,7 +663,7 @@ export class SpaceModel {
                             .andWhere((query2) => {
                                 query2
                                     .whereNotNull(
-                                        `${SpaceShareTableName}.user_id`,
+                                        `${SpaceUserAccessTableName}.user_uuid`,
                                     )
                                     .orWhere(
                                         `${ProjectMembershipsTableName}.role`,
@@ -686,9 +687,11 @@ export class SpaceModel {
                 `${UserTableName}.first_name`,
                 `${UserTableName}.last_name`,
                 `${EmailTableName}.email`,
+                `${SpaceTableName}.is_private`,
                 `${ProjectMembershipsTableName}.role`,
                 `${OrganizationMembershipsTableName}.role`,
-                `${SpaceShareTableName}.user_id`,
+                `${SpaceUserAccessTableName}.user_uuid`,
+                `${SpaceUserAccessTableName}.space_role`,
             )
             .select<
                 {
@@ -696,6 +699,8 @@ export class SpaceModel {
                     first_name: string;
                     last_name: string;
                     email: string;
+                    is_private: boolean;
+                    space_role: SpaceMemberRole;
                     user_with_direct_access: boolean;
                     project_role: ProjectMemberRole | null;
                     organization_role: OrganizationMemberRole;
@@ -706,8 +711,10 @@ export class SpaceModel {
                 `users.first_name`,
                 `users.last_name`,
                 `emails.email`,
+                `spaces.is_private`,
+                `space_user_access.space_role`,
                 this.database.raw(
-                    `CASE WHEN ${SpaceShareTableName}.user_id IS NULL THEN false ELSE true end as user_with_direct_access`,
+                    `CASE WHEN ${SpaceUserAccessTableName}.user_uuid IS NULL THEN false ELSE true end as user_with_direct_access`,
                 ),
                 `${ProjectMembershipsTableName}.role as project_role`,
                 `${OrganizationMembershipsTableName}.role as organization_role`,
@@ -724,6 +731,8 @@ export class SpaceModel {
                     first_name,
                     last_name,
                     email,
+                    is_private,
+                    space_role,
                     user_with_direct_access,
                     project_role,
                     organization_role,
@@ -756,6 +765,19 @@ export class SpaceModel {
                 if (!highestRole) {
                     return acc;
                 }
+
+                let spaceRole;
+
+                if (highestRole.role === ProjectMemberRole.ADMIN) {
+                    spaceRole = SpaceMemberRole.ADMIN;
+                } else if (user_with_direct_access) {
+                    spaceRole = space_role;
+                } else if (!is_private && !user_with_direct_access) {
+                    spaceRole = convertProjectRoleToSpaceRole(highestRole.role);
+                } else {
+                    return acc;
+                }
+
                 return [
                     ...acc,
                     {
@@ -763,7 +785,7 @@ export class SpaceModel {
                         firstName: first_name,
                         lastName: last_name,
                         email,
-                        role: convertProjectRoleToSpaceRole(highestRole.role),
+                        role: spaceRole,
                         hasDirectAccess: !!user_with_direct_access,
                         inheritedRole: highestRole.role,
                         inheritedFrom: highestRole.type,
@@ -946,7 +968,9 @@ export class SpaceModel {
         }));
     }
 
-    async getSpaceSummary(spaceUuid: string): Promise<SpaceSummary> {
+    async getSpaceSummary(
+        spaceUuid: string,
+    ): Promise<Omit<SpaceSummary, 'userAccess'>> {
         return wrapOtelSpan('SpaceModel.getSpaceSummary', {}, async () => {
             const [space] = await this.find({ spaceUuid });
             if (space === undefined)
@@ -962,10 +986,7 @@ export class SpaceModel {
     ): Promise<
         Map<
             string,
-            Pick<
-                SpaceSummary | Space,
-                'isPrivate' | 'access' | 'organizationUuid' | 'projectUuid'
-            >
+            Pick<SpaceSummary, 'isPrivate' | 'organizationUuid' | 'projectUuid'>
         >
     > {
         const spaces = await this.database('spaces')
@@ -975,11 +996,15 @@ export class SpaceModel {
                 'organizations.organization_id',
                 'projects.organization_id',
             )
-            .leftJoin('space_share', 'space_share.space_id', 'spaces.space_id')
+            .leftJoin(
+                'space_user_access',
+                'space_user_access.space_uuid',
+                'spaces.space_uuid',
+            )
             .leftJoin(
                 'users as shared_with',
-                'space_share.user_id',
-                'shared_with.user_id',
+                'space_user_access.user_uuid',
+                'shared_with.user_uuid',
             )
             .whereIn('spaces.space_uuid', spaceUuids)
             .select({
@@ -1003,7 +1028,6 @@ export class SpaceModel {
                 organizationUuid: space.organizationUuid,
                 projectUuid: space.projectUuid,
                 isPrivate: space.isPrivate,
-                access: space.access,
             });
         });
 
@@ -1075,21 +1099,18 @@ export class SpaceModel {
         return this.getFullSpace(spaceUuid);
     }
 
-    async addSpaceAccess(spaceUuid: string, userUuid: string): Promise<void> {
-        const [space] = await this.database('spaces')
-            .select('space_id')
-            .where('space_uuid', spaceUuid);
-
-        const [user] = await this.database('users')
-            .select('user_id')
-            .where('user_uuid', userUuid);
-
-        await this.database(SpaceShareTableName)
+    async addSpaceAccess(
+        spaceUuid: string,
+        userUuid: string,
+        spaceRole: SpaceMemberRole,
+    ): Promise<void> {
+        await this.database(SpaceUserAccessTableName)
             .insert({
-                space_id: space.space_id,
-                user_id: user.user_id,
+                space_uuid: spaceUuid,
+                user_uuid: userUuid,
+                space_role: spaceRole,
             })
-            .onConflict(['user_id', 'space_id'])
+            .onConflict(['user_uuid', 'space_uuid'])
             .merge();
     }
 
@@ -1097,32 +1118,9 @@ export class SpaceModel {
         spaceUuid: string,
         userUuid: string,
     ): Promise<void> {
-        const [space] = await this.database('spaces')
-            .select('space_id')
-            .where('space_uuid', spaceUuid);
-
-        const [user] = await this.database('users')
-            .select('user_id')
-            .where('user_uuid', userUuid);
-
-        await this.database(SpaceShareTableName)
-            .where('space_id', space.space_id)
-            .andWhere('user_id', user.user_id)
-            .delete();
-    }
-
-    async clearSpaceAccess(spaceUuid: string, userUuid: string): Promise<void> {
-        const [space] = await this.database('spaces')
-            .select('space_id')
-            .where('space_uuid', spaceUuid);
-
-        const [user] = await this.database('users')
-            .select('user_id')
-            .where('user_uuid', userUuid);
-
-        await this.database('space_share')
-            .where('space_id', space.space_id)
-            .andWhereNot('user_id', user.user_id)
+        await this.database(SpaceUserAccessTableName)
+            .where('space_uuid', spaceUuid)
+            .andWhere('user_uuid', userUuid)
             .delete();
     }
 }