+++ packages/frontend/src/components/LightdashVisualization/VisualizationBigNumberConfig.tsx
@@ -1,4 +1,8 @@
-import { ChartType, type ItemsMap } from '@lightdash/common';
+import {
+    ChartType,
+    type ItemsMap,
+    type TableCalculationMetadata,
+} from '@lightdash/common';
 import { useEffect, type FC } from 'react';
 import useBigNumberConfig from '../../hooks/useBigNumberConfig';
 import {
@@ -20,6 +24,7 @@ export const isBigNumberVisualizationConfig = (
 type VisualizationBigNumberConfigProps =
     VisualizationConfigCommon<VisualizationConfigBigNumber> & {
         itemsMap: ItemsMap | undefined;
+        tableCalculationsMetadata?: TableCalculationMetadata[];
     };
 
 const VisualizationBigNumberConfig: FC<VisualizationBigNumberConfigProps> = ({
@@ -28,11 +33,13 @@ const VisualizationBigNumberConfig: FC<VisualizationBigNumberConfigProps> = ({
     initialChartConfig,
     onChartConfigChange,
     children,
+    tableCalculationsMetadata,
 }) => {
     const bigNumberConfig = useBigNumberConfig(
         initialChartConfig,
         resultsData,
         itemsMap,
+        tableCalculationsMetadata,
     );
 
     useEffect(() => {