+++ packages/frontend/src/components/common/ResourceView/ResourceActionMenu.tsx
@@ -22,6 +22,7 @@ import { Fragment, useMemo, type FC } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import { useSpaceSummaries } from '../../../hooks/useSpaces';
 import { useApp } from '../../../providers/AppProvider';
+import { Can } from '../Authorization';
 import MantineIcon from '../MantineIcon';
 import {
     ResourceViewItemAction,
@@ -48,7 +49,7 @@ enum SpaceType {
 
 const SpaceTypeLabels = {
     [SpaceType.SharedWithMe]: 'Shared with me',
-    [SpaceType.AdminContentView]: 'Admin content view',
+    [SpaceType.AdminContentView]: 'Public content view',
 };
 
 const ResourceViewActionMenu: FC<ResourceViewActionMenuProps> = ({
@@ -63,57 +64,87 @@ const ResourceViewActionMenu: FC<ResourceViewActionMenuProps> = ({
     const location = useLocation();
     const { projectUuid } = useParams<{ projectUuid: string }>();
     const organizationUuid = user.data?.organizationUuid;
-    const { data: spaces = [] } = useSpaceSummaries(projectUuid, true);
+    const { data: spaces = [] } = useSpaceSummaries(projectUuid, true, {});
     const isPinned = !!item.data.pinnedListUuid;
     const isDashboardPage = location.pathname.includes('/dashboards');
 
-    const spacesSharedWithMe = useMemo(() => {
-        return spaces.filter((space) => {
-            return user.data && space.access.includes(user.data.userUuid);
+    const spacesByType = useMemo(() => {
+        const spacesUserCanCreateIn = spaces.filter((space) => {
+            return user.data?.ability?.can(
+                'create',
+                subject('SavedChart', {
+                    ...space,
+                    access: space.userAccess ? [space.userAccess] : [],
+                }),
+            );
         });
-    }, [spaces, user.data]);
-
-    const spacesAdminsCanSee = useMemo(() => {
-        return spaces.filter((space) => {
+        const spacesSharedWithMe = spacesUserCanCreateIn.filter((space) => {
+            return user.data && space.userAccess?.hasDirectAccess;
+        });
+        const spacesAdminsCanSee = spacesUserCanCreateIn.filter((space) => {
             return (
                 spacesSharedWithMe.find((s) => s.uuid === space.uuid) ===
                 undefined
             );
         });
-    }, [spaces, spacesSharedWithMe]);
-
-    const spacesByType = useMemo(() => {
         return {
             [SpaceType.SharedWithMe]: spacesSharedWithMe,
             [SpaceType.AdminContentView]: spacesAdminsCanSee,
         };
-    }, [spacesSharedWithMe, spacesAdminsCanSee]);
+    }, [spaces, user.data]);
 
     switch (item.type) {
-        case ResourceViewItemType.CHART:
-            if (user.data?.ability?.cannot('manage', 'SavedChart')) {
-                return null;
-            }
-            break;
-        case ResourceViewItemType.DASHBOARD:
-            if (user.data?.ability?.cannot('manage', 'Dashboard')) {
-                return null;
-            }
-            break;
-        case ResourceViewItemType.SPACE:
+        case ResourceViewItemType.CHART: {
+            const userAccess = spaces.find(
+                (space) => space.uuid === item.data.spaceUuid,
+            )?.userAccess;
             if (
                 user.data?.ability?.cannot(
                     'manage',
-                    subject('Space', {
-                        organizationUuid: item.data.organizationUuid,
-                        projectUuid,
-                        isPrivate: false,
+                    subject('SavedChart', {
+                        ...item.data,
+                        access: userAccess ? [userAccess] : [],
                     }),
                 )
             ) {
                 return null;
             }
             break;
+        }
+        case ResourceViewItemType.DASHBOARD: {
+            const userAccess = spaces.find(
+                (space) => space.uuid === item.data.spaceUuid,
+            )?.userAccess;
+            if (
+                user.data?.ability?.cannot(
+                    'manage',
+                    subject('Dashboard', {
+                        ...item.data,
+                        access: userAccess ? [userAccess] : [],
+                    }),
+                )
+            ) {
+                return null;
+            }
+            break;
+        }
+        case ResourceViewItemType.SPACE: {
+            const userAccess = spaces.find(
+                (space) => space.uuid === item.data.uuid,
+            )?.userAccess;
+            if (
+                user.data?.ability?.cannot(
+                    'manage',
+                    subject('Space', {
+                        ...item.data,
+                        access: userAccess ? [userAccess] : [],
+                    }),
+                )
+            ) {
+                return null;
+            }
+            break;
+        }
         default:
             return assertUnreachable(item, 'Resource type not supported');
     }
@@ -259,12 +290,13 @@ const ResourceViewActionMenu: FC<ResourceViewActionMenuProps> = ({
                                     SpaceType.AdminContentView,
                                 ].map((spaceType) => (
                                     <Fragment key={spaceType}>
-                                        {spacesByType[
-                                            SpaceType.AdminContentView
-                                        ].length > 0 ? (
+                                        {spacesByType[spaceType].length > 0 ? (
                                             <>
                                                 {spaceType ===
-                                                SpaceType.AdminContentView ? (
+                                                    SpaceType.AdminContentView &&
+                                                spacesByType[
+                                                    SpaceType.SharedWithMe
+                                                ].length > 0 ? (
                                                     <Menu.Divider />
                                                 ) : null}
 
@@ -322,26 +354,36 @@ const ResourceViewActionMenu: FC<ResourceViewActionMenuProps> = ({
                                     </Fragment>
                                 ))}
 
-                                {spaces.length > 0 ? <Menu.Divider /> : null}
-
-                                <Menu.Item
-                                    component="button"
-                                    role="menuitem"
-                                    icon={
-                                        <MantineIcon
-                                            icon={IconPlus}
-                                            size={18}
-                                        />
-                                    }
-                                    onClick={() => {
-                                        onAction({
-                                            type: ResourceViewItemAction.CREATE_SPACE,
-                                            item,
-                                        });
-                                    }}
+                                <Can
+                                    I="create"
+                                    this={subject('Space', {
+                                        organizationUuid:
+                                            user.data?.organizationUuid,
+                                        projectUuid,
+                                    })}
                                 >
-                                    Create new space
-                                </Menu.Item>
+                                    {spaces.length > 0 ? (
+                                        <Menu.Divider />
+                                    ) : null}
+                                    <Menu.Item
+                                        component="button"
+                                        role="menuitem"
+                                        icon={
+                                            <MantineIcon
+                                                icon={IconPlus}
+                                                size={18}
+                                            />
+                                        }
+                                        onClick={() => {
+                                            onAction({
+                                                type: ResourceViewItemAction.CREATE_SPACE,
+                                                item,
+                                            });
+                                        }}
+                                    >
+                                        Create new space
+                                    </Menu.Item>
+                                </Can>
                             </Menu.Dropdown>
                         </Menu>
                     </>